django/requirements.txt
asgiref==3.5.0
boto3==1.21.26
botocore==1.24.26
Django==4.0.3
django-anymail==8.5
django-crispy-forms==1.14.0
django-json-tag==0.0.1
django-storages==1.12.2
jmespath==1.0.0
password-validator==1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
